requirements.txt
asgiref==3.7.2
Django==3.2
django-filter==23.4
django-model-utils==4.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-secure-token==1.1.0
mysqlclient==2.2.0
phonenumbers==8.13.26
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
